Improve problem 1.3's probability convergence
The original approach would not ensure that we are within some precision of the specified probability. So if the probability of the predicate was 0.000001 we could stop when we get to 0.001 and think our answer is right. The way this works in this PR is that first the number of decimal places is calculated and then we iterate until our predicate seems to be within 2 decimal points of that for a few iterations.