Add a tool to explore 'stepping' (#115)
Currently when calculating the next composite for a prime not divisible by smaller primes we do trial division over all smaller primes. This PR explores what it would look like to keep track of "steps" that could be used to cycle between to hit the composites without having to do trial division (at least for small primes). It looks like it is feasible to do this for primes up to 19 or so.