Add functions for computing gaps and applied gaps (#244)
The gaps and applied gaps came up when counting twin primes in the RRS primorial sets. For RRS(1) = {1} using the naive approach of checking if 1 and 1 + 2 are twin primes results in it not counting {3, 5} as a twin prime. But all the other cases are fine. That's because for that first case the RRS set is not counting the desired quantity, which is the concept of applying the gaps to the RRS which gives {3} for the first case which does give (3, 5) as a twin prime. This is because primorial(1) < prime(1) -> 2 < 3 but for all greater numbers i > 1, primorial(i) > prime(i).

This just explains why when counting properties of 2s using just RRS's, a special case on 1 is needed: it's simpler often to work with RRSs because they are simpler to define but really what's desired is working with applied gaps to RRSs which are more complicated to define.