Explicitly skip the single un-added "child" for rrs() and children() (#249)
I'm not sure if this is faster but it's certainly more clear what is going on: skipping a single value in 0 <= k < prime(i). This also makes it more clear why the sizes of the reduced residue systems is the product of (p - 1) for the given primes since it's the number of values added for each residue in the earlier system.

See https://en.wikipedia.org/wiki/Modular_multiplicative_inverse for more on multiplicative inverses.
And https://docs.python.org/3/library/functions.html#pow for the documentation on pow https://bugs.python.org/issue36027 for the feature request to support it.

Also https://oeis.org/A062347 is neat because it is what the modular inverses are calculated for (primorial(i - 1) % prime(i)).