Optimize min_prime_descendants.py a bit (#262)
This makes it so fast that there is no benefit to outputting the progress.

This is done by caching prime and primorial which makes sense because they are rarely called for large values.

Profiled via
$ python -m cProfile -o a.profile -s cumulative fun/sieve/min_prime_descendants.py
$ python -m pstats a.profile
% sort cumtime
% stats

Now primality testing is using most of the time. It looks like the next thing to consider caching is the multiplicative inverse.